+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Further fixes for Thrashers.
 - Yak Pants should no longer cause a game crash with newer versions of Environmental.
 - Kinetic damage should no longer cause a server crash.
+- Ender Torches (Endergetic Expansion) now give the correct particles.
 
 ## [0.1.4] - 2024-04-18